requirements.txt
fastapi
uvicorn
pillow
opencv-python
torch
torchvision
torchaudio
python-multipart
requests
python-multipart
easyocr
fastapi
insightface
onnxruntime       # CPU version (remove if using GPU)
numpy
opencv-python-headless  # Headless version for servers (no GUI)
